@@ -1,8 +1,8 @@
 /// Benchmarks related to commit creation.
 /// We're measuring the impact of different parameters on the runtime.
 use criterion::{
-    async_executor::AsyncStdExecutor as FuturesExecutor, black_box, criterion_group, criterion_main, BatchSize,
-    Criterion,
+    BatchSize, Criterion, async_executor::AsyncStdExecutor as FuturesExecutor, black_box, criterion_group,
+    criterion_main,
 };
 
 use crate::utils::*;
