@@ -1,6 +1,6 @@
 use criterion::{
-    async_executor::AsyncStdExecutor as FuturesExecutor, black_box, criterion_group, criterion_main, BatchSize,
-    Criterion,
+    BatchSize, Criterion, async_executor::AsyncStdExecutor as FuturesExecutor, black_box, criterion_group,
+    criterion_main,
 };
 use rand::distributions::{Alphanumeric, DistString};
 
