@@ -2,10 +2,10 @@
 
 use super::Result;
 use crate::{
+    RecursiveError,
     context::CentralContext,
     mls,
     prelude::{Client, MlsCentral, MlsCredentialType},
-    RecursiveError,
 };
 use openmls_traits::types::SignatureScheme;
 
@@ -49,7 +49,7 @@ impl Client {
 
 #[cfg(test)]
 mod tests {
-    use crate::{e2e_identity::error::Error, mls, prelude::MlsCredentialType, test_utils::*, RecursiveError};
+    use crate::{RecursiveError, e2e_identity::error::Error, mls, prelude::MlsCredentialType, test_utils::*};
     use openmls_traits::types::SignatureScheme;
     use wasm_bindgen_test::*;
 
