@@ -14,15 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
 
-use openmls::prelude::{hash_ref::ProposalRef, KeyPackage};
+use openmls::prelude::{KeyPackage, hash_ref::ProposalRef};
 
 use mls_crypto_provider::MlsCryptoProvider;
 
 use super::{Error, Result};
 use crate::{
+    RecursiveError,
     mls::{ClientId, ConversationId, MlsConversation},
     prelude::{Client, MlsProposalBundle},
-    RecursiveError,
 };
 
 use crate::context::CentralContext;
